+++ src/main/java/roomescape/reservation/controller/ReservationController.java
@@ -11,7 +11,7 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 import roomescape.reservation.Reservation;
-import roomescape.reservation.ReservationService;
+import roomescape.reservation.service.ReservationService;
 import roomescape.reservation.dto.ReservationRequest;
 import roomescape.reservation.dto.ReservationResponse;
