+++ src/main/java/roomescape/theme/controller/ThemeController.java
@@ -9,7 +9,7 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import roomescape.theme.Theme;
-import roomescape.theme.ThemeService;
+import roomescape.theme.service.ThemeService;
 import roomescape.theme.dto.ThemeResponse;
 
 @RestController
